Added changes

+++ src/main/java/org/game/java/GameObject.java
@@ -7,12 +7,12 @@ import java.awt.*;
  */
 public abstract class GameObject {
 
-    protected int x, y;
+    protected float x, y;
     protected ID id;
-    protected int velX, velY;
+    protected float velX, velY;
 
 
-    public GameObject(int x, int y, ID id) {
+    public GameObject(float x, float y, ID id) {
         this.x = x;
         this.y = y;
         this.id = id;
@@ -32,11 +32,11 @@ public abstract class GameObject {
         this.y = y;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
@@ -56,11 +56,11 @@ public abstract class GameObject {
         this.velY = velY;
     }
 
-    public int getVelX() {
+    public float getVelX() {
         return velX;
     }
 
-    public int getVelY() {
+    public float getVelY() {
         return velY;
     }
