+++ src/main/java/org/game/java/HUD.java
@@ -7,9 +7,9 @@ import java.awt.*;
  */
 public class HUD {
 
-    public static int HEALTH = 100;
+    public static float HEALTH = 100;
 
-    private int greenVariable = 255;
+    private float greenVariable = 255;
 
     private int score = 0;
     private int level = 1;
@@ -24,8 +24,8 @@ public class HUD {
     public void render(Graphics g) {
         g.setColor(Color.GRAY);
         g.fillRect(15, 15, 200, 32);
-        g.setColor(new Color(75, greenVariable, 0));
-        g.fillRect(15, 15, HEALTH * 2, 32);
+        g.setColor(new Color(75, (int)greenVariable, 0));
+        g.fillRect(15, 15, (int)HEALTH * 2, 32);
         g.setColor(Color.WHITE);
         g.drawRect(15, 15, 200, 32);
